Move controller into separate package
